personalized exception to block deletion from incomplete lists

+++ app/usecases/task_list.py
@@ -1,5 +1,6 @@
-from app.domain.utils import extract_non_null_fields
 from app.domain.entities import TaskList
+from app.exceptions import TaskListDeletionError
+from app.domain.utils import extract_non_null_fields
 from app.domain.interfaces import ITaskListRepository
 
 
@@ -23,4 +24,7 @@ class TaskListUsecase:
         return await self.task_list_repository.update(task_list_id, update_dict)
     
     async def delete_list(self, task_list_id: int) -> bool:
+        task_list = await self.task_list_repository.get(task_list_id)
+        if any(not task.is_completed for task in task_list.tasks):
+            raise TaskListDeletionError()
         return await self.task_list_repository.delete(task_list_id)
